Set the language and section hashcode to the SFX files when exporting.

+++ Solutions/EuroSound Editor/Forms/SFX Form/SFXForm_TestSfx.cs
@@ -79,7 +79,7 @@ namespace sb_editor.Forms
                     DirectoryInfo musXFolder = Directory.CreateDirectory(outputFolder);
 
                     //Build File
-                    MusXBuild_Soundbank.BuildSoundbankFile(sfxTempFile, sifTempFile, sbfTempFile, string.Empty, Path.Combine(musXFolder.FullName, fileName), CommonFunctions.GetPlatformLabel("PC"), hashCode, false);
+                    MusXBuild_Soundbank.BuildSoundbankFile(sfxTempFile, sifTempFile, sbfTempFile, string.Empty, Path.Combine(musXFolder.FullName, fileName), CommonFunctions.GetPlatformLabel("PC"), (uint)hashCode, false);
                 }
             }
         }
